+++ src/app/app.component.ts
@@ -1,16 +1,15 @@
 import { JsonPipe } from '@angular/common';
-import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup, ReactiveFormsModule } from '@angular/forms';
+import { Component } from '@angular/core';
+import { FormBuilder, ReactiveFormsModule } from '@angular/forms';
 import {
   NgxCurrencyDirective,
   NgxCurrencyInputMode,
 } from '../../projects/ngx-currency/src/public-api';
 
 @Component({
-  standalone: true,
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss'],
+  styleUrl: './app.component.scss',
   imports: [ReactiveFormsModule, JsonPipe, NgxCurrencyDirective],
 })
 export class AppComponent implements OnInit {
@@ -28,15 +27,13 @@ export class AppComponent implements OnInit {
   };
   ngxCurrencyInputMode = NgxCurrencyInputMode;
 
-  constructor(private formBuilder: FormBuilder) {
-    this.form = this.formBuilder.group({
-      value: 0,
-      inputMode: this.ngxCurrencyOptions.inputMode,
-    });
-  }
+  form = this.formBuilder.nonNullable.group({
+    value: 0,
+    inputMode: this.ngxCurrencyOptions.inputMode,
+  });
 
-  ngOnInit() {
-    this.form.controls['inputMode'].valueChanges.subscribe((val) => {
+  constructor(private readonly formBuilder: FormBuilder) {
+    this.form.controls.inputMode.valueChanges.subscribe(val => {
       this.ngxCurrencyOptions.inputMode = val;
 
       // Clear the value input when the input mode is changed.container